requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
requests==2.31.0
celery[redis]==5.2.7
django-celery-beat
django-cors-headers